globals support (a few old sketches have a null ptr exception; will fix that soon)

+++ src/main/java/sketch/compiler/main/PlatformLocalization.java
@@ -187,19 +187,17 @@ public class PlatformLocalization {
         String canonicalName = null;
         try {
             canonicalName = path.getCanonicalPath();
-        } catch (IOException e1) {
-            e1.printStackTrace();
-        }
-        if (path.exists()) {
-            return canonicalName;
-        } else {
-            try {
+            if (path.exists() && (path.length() == fileIn.available())) {
+                return canonicalName;
+            } else {
                 FileOutputStream fileOut = new FileOutputStream(path);
                 byte[] buffer = new byte[8192];
                 int len;
                 while ((len = fileIn.read(buffer)) > 0) {
                     fileOut.write(buffer, 0, len);
                 }
+                fileOut.flush();
+                fileOut.close();
                 assert (fileIn.available() == 0) : "didn't read all of file";
                 if (!isWin()) {
                     Process proc =
@@ -207,9 +205,9 @@ public class PlatformLocalization {
                     assert (proc.waitFor() == 0) : "couldn't make cegis executable";
                 }
                 return canonicalName;
-            } catch (Exception e) {
-                System.err.println("couldn't extract cegis binary; " + e);
             }
+        } catch (Exception e) {
+            System.err.println("couldn't extract cegis binary; " + e);
         }
         return null;
     }
